chore(components): move large logo from ui to comps
