+++ packages/nextjs/components/index.ts
@@ -1,2 +1,3 @@
 export { Header } from "./Header";
 export { Footer } from "./Footer";
+export { LargeLogo } from "./LargeLogo";
