+++ packages/nextjs/ui/index.ts
@@ -1,7 +1,6 @@
 export { Button } from "./Button/Button";
 export { Column } from "./Column";
 export { Input } from "./Input/Input";
-export { LargeLogo } from "./LargeLogo";
 export { Loader } from "./Loader";
 export { Text } from "./Text/Text";
 export { Row } from "./Row";
